+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-## 1.35.0 (May 22, 2023)
+## 1.35.0 (May 26, 2023)
+
+FEATURES
+
+* **New Resource**: videooptimizationpacingaction 
+* **New Resource**: videooptimizationpacingpolicy 
 
 BUG FIXES
 
+* **citrixadc_ntpsync**: Updated create operation to call appropriate API call.
 * **citrixadc_appfwprofile**: Updated the read function to not to set values for some attributes that are not recieved from the NetScaler.
+
 
 ## 1.34.0 (May 05, 2023)